Skip newrelic install for non-amd64 arch

+++ build/scripts/installNewRelic.sh
@@ -1,18 +1,22 @@
 #!/usr/bin/env sh
 set -e
 
-# Get latest download URI.
-NEWRELIC_DOWNLOAD_FILE=$(curl -s https://download.newrelic.com/php_agent/release/ | sed -rn 's/.*(newrelic-php5-[0-9.]+-linux-musl\.tar\.gz).*/\1/p')
-NEWRELIC_DOWNLOAD_URI="https://download.newrelic.com/php_agent/release/$NEWRELIC_DOWNLOAD_FILE"
+if [ "$TARGETPLATFORM" = "linux/amd64" ]; then
+  # Get latest download URI.
+  NEWRELIC_DOWNLOAD_FILE=$(curl -s https://download.newrelic.com/php_agent/release/ | sed -rn 's/.*(newrelic-php5-[0-9.]+-linux-musl\.tar\.gz).*/\1/p')
+  NEWRELIC_DOWNLOAD_URI="https://download.newrelic.com/php_agent/release/$NEWRELIC_DOWNLOAD_FILE"
 
-# Install Package.
-mkdir -p /opt/newrelic
-cd /opt/newrelic
-wget "$NEWRELIC_DOWNLOAD_URI" -O "/opt/newrelic/$NEWRELIC_DOWNLOAD_FILE"
-tar -zxf "$NEWRELIC_DOWNLOAD_FILE"
-cd /opt/newrelic/newrelic-php5-*
-NR_INSTALL_SILENT="true" sh newrelic-install install
+  # Install Package.
+  mkdir -p /opt/newrelic
+  cd /opt/newrelic
+  wget "$NEWRELIC_DOWNLOAD_URI" -O "/opt/newrelic/$NEWRELIC_DOWNLOAD_FILE"
+  tar -zxf "$NEWRELIC_DOWNLOAD_FILE"
+  cd /opt/newrelic/newrelic-php5-*
+  NR_INSTALL_SILENT="true" sh newrelic-install install
 
-# Tidy up temporary files.
-rm -f "/opt/newrelic/$NEWRELIC_DOWNLOAD_FILE"
-rm -f /tmp/nrinstall-*
+  # Tidy up temporary files.
+  rm -f "/opt/newrelic/$NEWRELIC_DOWNLOAD_FILE"
+  rm -f /tmp/nrinstall-*
+else
+  echo "New Relic not supported on amd64 architecture. Skipping install..."
+fi
